Odděl náhradu grafémů
Náhrada grafémů za fonémy je činnost nesouvisející se samotnou morfologickou segmentací. Jedná se o oddělený proces, nyní vyčleněný do vlastního skriptu.
